# feat(plugin-chart-pivot-table): support series limit