### SUMMARY
Support series limit along with row limit in Pivot Table v2.

Migration benchmark results:
10+: 0.38 s
100+: 0.25 s
1000+: 0.24 s
10000+: 0.36 s
100000+: 0.46 s

### BEFORE/AFTER SCREENSHOTS OR ANIMATED GIF
Before:
![image](https://user-images.githubusercontent.com/15073128/146569777-8806be44-09e3-4125-aa04-7ff506f05538.png)

After:
![image](https://user-images.githubusercontent.com/15073128/146569631-6d18e4cc-02fc-4e0b-92d0-a8c72c0baf31.png)

### TESTING INSTRUCTIONS
1. Create a Pivot Table v2 chart
2. Set a series limit and series limit metric
3. Click run and go to View Query
4. Verify that the query contains a JOIN subquery ordered by selected series limit metric and limit equal to series limit
5. Verify that the query limit is equal to Cell limit and is ordereb by series limit metric.
6. Remove series limit metric and rerun the query
7. Verify that the query and JOIN subquery are ordered by the first selected metric in Metrics control

### ADDITIONAL INFORMATION
<!--- Check any relevant boxes with "x" -->
<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->
- [x] Has associated issue: closes #16901, #13562
- [ ] Required feature flags:
- [ ] Changes UI
- [x] Includes DB Migration (follow approval process in [SIP-59](https://github.com/apache/superset/issues/13351))
  - [x] Migration is atomic, supports rollback & is backwards-compatible
  - [x] Confirm DB migration upgrade and downgrade tested
  - [x] Runtime estimates and downtime expectations provided
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API

CC @rusackas @graceguo-supercat