@zhaoyongjie 2021-12-21
Can't use custom metrics in sort control.

![image](https://user-images.githubusercontent.com/2016594/146891900-7a2e97fa-0da5-4228-bc04-00ebc8971e35.png)

![image](https://user-images.githubusercontent.com/2016594/146891801-01207eb2-7e7a-4972-83dc-3e7ee2d5864c.png)