@kgabryje 2021-12-21
> Can't use custom metrics in sort control.

Thank you @zhaoyongjie, well spotted! Issue fixed. Can you retest?

![image](https://user-images.githubusercontent.com/15073128/146912311-27b2ae16-3998-4d45-9b15-331cf8dc0742.png)